[Chore] 익스텐션 정리
