+++ DaangnMarket-iOS/DaangnMarket-iOS/Utils/Adjusted+.swift
